Generate a bunch of tables
For each language, up to three tables are created: all works, long works,
and short works.  Each of these tables shows how many such works there are,
how many have been read, and what percentage that is, as well as the
corresponding values for the number of pages (which is proportional to the
number of characters) and number of words.